control: allow to disable server-port for outside rtsp udp streams.
Add configuration flag per stream that allows to disable rtsp udp server
data port. This is required to similate rtsp cameras/servers that do not
provide that functionality.
When server-port is not provided client does not send any packet during
initial conversation to that UDP port and "firewall" punching is not
happening. This results in a behaviour that if "stream reader" runs
inside container on non-host network UDP rtsp will not work unless ports
are explicitly forwarded.
This setting is added especially in order to recreate such behaviour.